% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_list_calcs.R
\name{model.list}
\alias{model.list}
\title{Create a list of all possible models from a set of covariates and a response variable}
\usage{
model.list(covariates, response)
}
\arguments{
\item{response}{A string with the response variable followed by a ~ and the intercept
(usually 1), e.g. "Mean_Abundance_perTrapnight~1"}

\item{regressors}{Vector of strings representing each covariate in a gam model
including k value and outter wraper (s, ti, te etc.)}
}
\value{
A list of formulas for all possible combinations of models and a matrix
using logical values to describe which terms are included in each of the model formulas.
}
\description{
Create a list of all possible models from a set of covariates and a response variable.
Generally, this function is only useful when the resulting object is passed to
dev.expl.mgcv::dev.expl.calcs.
}
\examples{
#from mgcv documentation
library(mgcv)
library(dev.expl.mgcv)
set.seed(2) ## simulate some data...
dat <- gamSim(1,n=400,dist="normal",scale=2)
b <- gam(y~1+x0+s(x0, k=4)+ti(x2, x3, k=4),data=dat)
model.list.output<-model.list(covariates = c("x0","s(x0, k=4)",
"ti(x2, x3, k=4)"), response="y~1")
}
